Add source code for status feature

+++ social media/src/main/java/com/codegym/socialmedia/model/social_action/FriendshipId.java
@@ -1,3 +1,4 @@
+// FriendshipId.java - Fixed version
 package com.codegym.socialmedia.model.social_action;
 
 import jakarta.persistence.Embeddable;
@@ -6,12 +7,13 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Embeddable
 public class FriendshipId implements Serializable {
-    private Integer requesterId;
-    private Integer addresseeId;
-
+    // FIX: Thay đổi từ Integer thành Long để match với User.id
+    private Long requesterId;
+    private Long addresseeId;
 }
